+++ ThreeJs4Net.Tests/Examples/Jsm/Utils/BufferGeometryUtilsTests.cs
@@ -1,9 +1,6 @@
 ﻿using System;
-using System.Runtime.Intrinsics.X86;
 using ThreeJs4Net.Core;
-using ThreeJs4Net.Examples.Jsm.Math;
 using ThreeJs4Net.Examples.Jsm.Utils;
-using ThreeJs4Net.Math;
 using Xunit;
 
 namespace ThreeJs4Net.Tests.Examples.Jsm.Utils
@@ -13,15 +10,15 @@ namespace ThreeJs4Net.Tests.Examples.Jsm.Utils
         [Fact()]
         public void MergeBufferAttributesBasic()
         {
-            var array1 = new float[] { 1, 2, 3, 4 };
-            var attr1 = new BufferAttribute<float>(array1, 2, false);
+            var array1 = new double[] { 1, 2, 3, 4 };
+            var attr1 = new BufferAttribute<double>(array1, 2, false);
 
-            var array2 = new float[] { 5, 6, 7, 8 };
-            var attr2 = new BufferAttribute<float>(array2, 2, false);
+            var array2 = new double[] { 5, 6, 7, 8 };
+            var attr2 = new BufferAttribute<double>(array2, 2, false);
 
-            var mergedAttr = BufferGeometryUtils.MergeBufferAttributes<float>(new[] { attr1, attr2 });
+            var mergedAttr = BufferGeometryUtils.MergeBufferAttributes<double>(new[] { attr1, attr2 });
 
-            Assert.Equal(mergedAttr.Array, new float[] { 1, 2, 3, 4, 5, 6, 7, 8 });
+            Assert.Equal(mergedAttr.Array, new double[] { 1, 2, 3, 4, 5, 6, 7, 8 });
             Assert.Equal(2, mergedAttr.ItemSize);
             Assert.False(mergedAttr.Normalized);
 
@@ -30,41 +27,41 @@ namespace ThreeJs4Net.Tests.Examples.Jsm.Utils
         [Fact()]
         public void MergeBufferAttributesInvalid()
         {
-            var array1 = new float[] { 1, 2, 3, 4 };
-            var attr1 = new BufferAttribute<float>(array1, 2, false);
+            var array1 = new double[] { 1, 2, 3, 4 };
+            var attr1 = new BufferAttribute<double>(array1, 2, false);
 
-            var array2 = new float[] { 5, 6, 7, 8 };
-            var attr2 = new BufferAttribute<float>(array2, 4, false);
+            var array2 = new double[] { 5, 6, 7, 8 };
+            var attr2 = new BufferAttribute<double>(array2, 4, false);
 
-            Assert.Throws<Exception>(() => BufferGeometryUtils.MergeBufferAttributes<float>(new[] { attr1, attr2 }));
+            Assert.Throws<Exception>(() => BufferGeometryUtils.MergeBufferAttributes<double>(new[] { attr1, attr2 }));
 
             attr2.ItemSize = 2;
             attr2.Normalized = true;
 
-            Assert.Throws<Exception>(() => BufferGeometryUtils.MergeBufferAttributes<float>(new[] { attr1, attr2 }));
+            Assert.Throws<Exception>(() => BufferGeometryUtils.MergeBufferAttributes<double>(new[] { attr1, attr2 }));
 
             attr2.Normalized = false;
 
-            Assert.NotNull(BufferGeometryUtils.MergeBufferAttributes<float>(new[] { attr1, attr2 }));
+            Assert.NotNull(BufferGeometryUtils.MergeBufferAttributes<double>(new[] { attr1, attr2 }));
         }
 
         [Fact()]
         public void MergeBufferGeometryBasic()
         {
             var geometry1 = new BufferGeometry();
-            geometry1.SetAttribute("position", new BufferAttribute<float>(new float[] { 1, 2, 3 }, 1, false));
+            geometry1.SetAttribute("position", new BufferAttribute<double>(new double[] { 1, 2, 3 }, 1, false));
             geometry1.SetAttribute("uint", new BufferAttribute<uint>(new uint[] { 1, 2, 3 }, 1, false));
 
             var geometry2 = new BufferGeometry();
-            geometry2.SetAttribute("position", new BufferAttribute<float>(new float[] { 4, 5, 6 }, 1, false));
+            geometry2.SetAttribute("position", new BufferAttribute<double>(new double[] { 4, 5, 6 }, 1, false));
             geometry2.SetAttribute("uint", new BufferAttribute<uint>(new uint[] { 1, 2, 3 }, 1, false));
 
             var mergedGeometry = BufferGeometryUtils.MergeBufferGeometries(new[] { geometry1, geometry2 });
 
             Assert.NotNull(mergedGeometry);
-            var position = mergedGeometry.GetAttribute<float>("position");
+            var position = mergedGeometry.GetAttribute<double>("position");
             var positions = position.Array;
-            Assert.Equal(positions, new float[] { 1, 2, 3, 4, 5, 6 });
+            Assert.Equal(positions, new double[] { 1, 2, 3, 4, 5, 6 });
             Assert.Equal(1, position.ItemSize);
         }
 
@@ -72,21 +69,21 @@ namespace ThreeJs4Net.Tests.Examples.Jsm.Utils
         public void MergeBufferGeometryIndexed()
         {
             var geometry1 = new BufferGeometry();
-            geometry1.SetAttribute("position", new BufferAttribute<float>(new float[] { 1, 2, 3 }, 1, false));
+            geometry1.SetAttribute("position", new BufferAttribute<double>(new double[] { 1, 2, 3 }, 1, false));
             geometry1.SetIndex(new BufferAttribute<uint>(new uint[] { 0, 1, 2, 2, 1, 0 }, 1, false));
 
             var geometry2 = new BufferGeometry();
-            geometry2.SetAttribute("position", new BufferAttribute<float>(new float[] { 4, 5, 6 }, 1, false));
+            geometry2.SetAttribute("position", new BufferAttribute<double>(new double[] { 4, 5, 6 }, 1, false));
             geometry2.SetIndex(new BufferAttribute<uint>(new uint[] { 0, 1, 2 }, 1, false));
 
             var mergedGeometry = BufferGeometryUtils.MergeBufferGeometries(new[] { geometry1, geometry2 });
 
             Assert.NotNull(mergedGeometry);
 
-            var position = mergedGeometry.GetAttribute<float>("position");
+            var position = mergedGeometry.GetAttribute<double>("position");
             var positions = position.Array;
 
-            Assert.Equal(positions, new float[] { 1, 2, 3, 4, 5, 6 });
+            Assert.Equal(positions, new double[] { 1, 2, 3, 4, 5, 6 });
             Assert.Equal(mergedGeometry.Index.Array, new uint[] { 0, 1, 2, 2, 1, 0, 3, 4, 5 });
             Assert.Equal(1, position.ItemSize);
 
@@ -96,11 +93,11 @@ namespace ThreeJs4Net.Tests.Examples.Jsm.Utils
         public void MergeBufferGeometryInvalid()
         {
             var geometry1 = new BufferGeometry();
-            geometry1.SetAttribute("position", new BufferAttribute<float>(new float[] { 1, 2, 3 }, 1, false));
+            geometry1.SetAttribute("position", new BufferAttribute<double>(new double[] { 1, 2, 3 }, 1, false));
             geometry1.SetIndex(new BufferAttribute<uint>(new uint[] { 0, 1, 2 }, 1, false));
 
             var geometry2 = new BufferGeometry();
-            geometry2.SetAttribute("position", new BufferAttribute<float>(new float[] { 4, 5, 6 }, 1, false));
+            geometry2.SetAttribute("position", new BufferAttribute<double>(new double[] { 4, 5, 6 }, 1, false));
 
             Assert.Throws<Exception>(() => BufferGeometryUtils.MergeBufferGeometries(new[] { geometry1, geometry2 }));
 
@@ -108,7 +105,7 @@ namespace ThreeJs4Net.Tests.Examples.Jsm.Utils
 
             Assert.NotNull(BufferGeometryUtils.MergeBufferGeometries(new[] { geometry1, geometry2 }));
 
-            geometry2.SetAttribute("foo", new BufferAttribute<float>(new float[] { 1, 2, 3 }, 1, false));
+            geometry2.SetAttribute("foo", new BufferAttribute<double>(new double[] { 1, 2, 3 }, 1, false));
 
             Assert.Throws<Exception>(() => BufferGeometryUtils.MergeBufferGeometries(new[] { geometry1, geometry2 }));
         }
